implemented webpack succesfully
