+++ index.html
@@ -7,14 +7,7 @@
     <title>Sabka Bazar</title>
     <!-- <base href="/"> -->
     <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Dosis:400,600,700">
-    <link type="text/css" rel="stylesheet" href="./shared/css/style.css">
     <script src="http://code.jquery.com/jquery-3.3.1.min.js" integrity="sha256-FgpCb/KJQlLNfOu91ta32o/NMZxltwRo8QtmkMRdAu8=" crossorigin="anonymous"></script>
-    
-    <script src="src/app/index.js"></script>
-    <script src="src/app/signup/signup.js"></script>
-    <script src="src/app/login/login.js"></script>
-    <script src="src/app/home/home.js"></script>
-    <script src="src/app/products/products.js"></script>
   </head>
   <body>
     <header class="header">
@@ -42,5 +35,12 @@
         <span>Copyright @ 2011-2019 Sabka Bazar Grocery Supplies Pvt Ltd</span>
       </div>
     </footer>
+    
+    <script src="js/home/home.js"></script>
+    <script src="js/login/login.js"></script>
+    <script src="js/signup/signup.js"></script>
+    <script src="js/products/products.js"></script>
+    <script src="js/main.js"></script>
+
   </body>
 </html>
