+++ webpack.config.js
@@ -17,7 +17,12 @@ module.exports = {
         template: './index.html'
     }),
     new CopyWebpackPlugin([
-        { from: './assets/images/', to: 'images' }
+        { from: './assets/images/', to: 'images' },
+        { from: './src/app/home', to: 'js/home' },
+        { from: './src/app/signup', to: 'js/signup' },
+        { from: './src/app/login', to: 'js/login' },
+        { from: './src/app/products', to: 'js/products'},
+        { from: './src/app', to: 'js'}
     ])
   ],
   module: {
